refactor(product-api): rename all methods
